Added basic skeleton code for doing validation against genstrategy rules.
