+++ lfr/notes.md
@@ -1,0 +1,29 @@
+#  Gen Strategys:
+
+
+## DropX Flow Validation Rules:
+
+### Rule 1 - 
+
+The first level of % should be mapping to a Droplet Generator
+
+### Rule 2 – 
+
+Any +-, distribute nodes before % should be in continuous flow
+
+### (Incorrect) Rule 3 - 
+
+Any Remetering (%) should require a droplet breakdown and regeneration
+
+### Rule 4 – 
+
+Distribute network post Metering stage should be mapped to different kinds of separator / selection/ storage networks
+
+### Rule 5 –
+
+If plus is shown between node that has % in pred and non % in pred, then its pico injection
+
+### Rule 6 – 
+
+If plus is sown between two nodes that has % in pred, then its droplet merging
+
