+++ tests/pylfr/e2e/compiler/test_mintgen.py
@@ -12,6 +12,10 @@ REF_DATA_FIG_FOLDER = TEST_CASES_FOLDER.joinpath("netlists")
 
 
 def test_dx1():
+
+    # Create output directories
+    TEST_OUTPATH.mkdir(parents=True, exist_ok=True)
+
     # dx1.lfr
     test_file = TEST_CASES_FOLDER.joinpath("dx1.lfr")
